remove observable from BettingChip
